% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatPathAddIn.R
\name{formatPathAddIn}
\alias{formatPathAddIn}
\title{Format File Path}
\usage{
formatPathAddIn()
}
\value{
Formatted file path using \code{\link[base]{file.path}}.
}
\description{
The add-in formats string file path into a call using \code{file.path}
  function.
}
\details{
The add-in breaks the path string, example"
  \code{"~/stuff/projects/r/something/else"} into \code{\link[base]{file.path}}
  call. The function can optionally call \code{\link[base]{normalizePath}} on the
  passed string and modify default \code{fsep} argument of the
  \code{\link[base]{file.path}}.
}
\references{
The add-in draws heavily on the code available in
  \url{https://github.com/rstudio/addinexamples}.
}
