% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_file_path_call.R
\name{create_file_path_call}
\alias{create_file_path_call}
\title{Create File Path Call}
\usage{
create_file_path_call(path_string, path_sep = .Platform$file.sep,
  mustWork = TRUE, normalize = TRUE)
}
\arguments{
\item{path_string}{Path string.}

\item{path_sep}{Path separator, defults to \code{.Platform$file.sep}.}

\item{mustWork}{As per \code{mustWork} in \code{\link[base]{normalizePath}}}

\item{normalize}{A logical, defaults to \code{TRUE} whether to wrap in
\code{\link[base]{normalizePath}}.}
}
\value{
A \code{\link[base]{file.path}} call.
}
\description{
The function returns \code{\link[base]{file.path}} function
  wrapping passed file path string.
}
\details{
The function is a key element of the \code{\link{formatPathAddIn}}
  and produces \code{\link[base]{file.path}} call.
}
\examples{
create_file_path_call(tempdir())

}
