% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatFilePath.R
\docType{package}
\name{reformatFilePath}
\alias{reformatFilePath}
\alias{reformatFilePath-package}
\title{RStudio Add-In Formatting File Path}
\description{
reformatFilePath RStudio Add-In formats file path string using
  \code{\link[base]{file.path}} call. The add-in can optionally add
  \code{\link[base]{normalizePath}} argument and control other arguments
  passed to \code{\link[base]{file.path}} and \code{\link[base]{normalizePath}}
  functions.
}
